algs/src/python/imagep/requirements.txt
pillow==9.0.1
numpy==1.22.3